payment description has been added

+++ app/src/main/java/com/leon/hamrah_abfa/fragments/dashboard/PaymentStats.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 
 public class PaymentStats {
     public String firstBillDate;
-    public float totalBills;
-    public float totalPayments;
-    public float unpayedBills;
-    public float payAverageTime;
+    public int totalBills;
+    public int totalPayments;
+    public int unpayedBills;
+    public int payAverageTime;
     public ArrayList<Integer> payDeadlineValues = new ArrayList<>();
     public ArrayList<String> payDeadlineKeys = new ArrayList<>();
 }
